fix: Improve URL handling and add error checks in attachment saving

+++ src/licitpy/countries/cl/utils.py
@@ -13,6 +13,7 @@ async def save_attachment(
 ) -> str:
     # If content is not provided, fetch it from the attachment
     if content is None:
+        # this triggers the download of the attachment content
         content = await attachment.content
 
     if not content:
@@ -20,8 +21,14 @@ async def save_attachment(
             f"Failed to download attachment: {attachment.name}"
         )
 
+    # If filename is not provided, use the attachment's name
     filename = filename or attachment.name
+
     full_path = os.path.join(path, filename)
+
+    # if folder does not exist , raise an error
+    if not os.path.exists(path):
+        raise FileNotFoundError(f"Directory does not exist: {path}")
 
     with open(full_path, "wb") as file:
         file.write(base64.b64decode(content))
